[TASK-1] CommonJS
Work done:
* CommonJS module
